Pin the Python protobuf dependency
The Python protobuf library was updated tonight in a backwards incompatible
way, and the Python bindings are no longer working for fresh builds. I tried
to upgrade the bindings but couldn't get the tests to pass. As a workaround,
this commit just pins the version of the Python protobuf package to something
less than 4.